pyproject.toml
[tool.poetry]
name = "todoappfullstack"
version = "0.1.0"
description = "A Todo fullstack app with sqlalchemy, fastapi and jinja"
authors = ["ikari2k <pluta.m@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.4.2"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pytest = "^8.2.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pytest-cov = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
